requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
coverage==7.5.4
exceptiongroup==1.2.1
Faker==26.0.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
tomli==2.0.1
urllib3==2.2.2
